refactor constructors
