+++ csr/constructors.py
@@ -1,0 +1,56 @@
+"""
+Numba-accessible constructors for CSRs.
+"""
+
+from .csr import CSR
+
+import numpy as np
+from numba import njit
+
+
+@njit
+def create_empty(nrows, ncols):
+    """
+    Create an empty CSR of the specified size.
+
+    .. note:: This function can be used from Numba.
+    """
+    rowptrs = np.zeros(nrows + 1, dtype=np.intc)
+    colinds = np.zeros(0, dtype=np.intc)
+    values = np.zeros(0)
+    return CSR(np.int32(nrows), np.int32(ncols), 0, rowptrs, colinds, True, values)
+
+
+@njit
+def create_novalues(nrows, ncols, nnz, rowptrs, colinds):
+    """
+    Create a CSR without values.
+    """
+    return CSR(nrows, ncols, nnz, rowptrs, colinds, False, np.zeros(0))
+
+
+@njit
+def create(nrows, ncols, nnz, rowptrs, colinds, values):
+    """
+    Create a CSR.
+    """
+    return CSR(nrows, ncols, nnz, rowptrs, colinds, True, values)
+
+
+@njit
+def create_from_sizes(nrows, ncols, sizes):
+    """
+    Create a CSR with uninitialized values and specified row sizes.
+
+    Args:
+        nrows(int): the number of rows
+        ncols(int): the number of columns
+        sizes(numpyp.ndarray): the number of nonzero values in each row
+    """
+    nnz = np.sum(sizes)
+    rowptrs = np.zeros(nrows + 1, dtype=sizes.dtype)
+    for i in range(nrows):
+        rowptrs[i + 1] = rowptrs[i] + sizes[i]
+    colinds = np.full(nnz, -1, dtype=np.intc)
+    values = np.full(nnz, np.nan)
+    return CSR(nrows, ncols, nnz, rowptrs, colinds, True, values)
